loop

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -8,7 +8,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-unsigned int i;
+long unsigned int i;
  
 if (array || action)
 {
